\name{getdata98}
\alias{getdata98}
\title{obtiene datos de la base de kpis del servidor 98}
\usage{
  getdata98 (variabale, mes, n)
}
\arguments{
  \item{variable}{variable de los kPIs a consultar. Requireda.}\item{mes}{AAAAMM mes a consultar. Requireda.} \item{n}{tamaño de la muestra. Requireda.}
}
\description{
  Esta función es una bomba, nos saltamos a seguridad HAHAHA
}

